Finale (#27)
* Rework interpolation for a hybrid approach, kind of like Starship but a little different

* Swap nabnut summer and winter (wrong filenames)

* Finish interpolation, adjust hud centering (again) and add enhancement for all honeycombs

* Fix Furnace Fun

* More furnace fun fixing and double health enhancement correction

* Fix final fight

* Fix cutscene timing for some select scenes

* Fix skipped jinjonator sequence at end of fight

* Fix sns demo pictures timing

* Correctly name and move a file

* Rename and move more files to make sense

* Darn header files

* Fix framebuffer flashes for SNS pictures

* Performance Phase 1: Move aux fbo to GPU

* Improve pause menu performance and quality

* Phase 1.5 - AuxFBO swapped over to standard LUS fbo

* Phase 2: Complete switch to GPU framebuffer and create Patches file, restore old decomp content in auxbuffer.c

* Clang-format

* Update libultraship

+++ src/LAIR/ff_manager.c
@@ -59,22 +59,14 @@ extern void func_802FACA4(enum item_e);
 extern void func_8028FA14(enum map_e, s32);
 extern void quizQuestionAskedBitfield_set(u32, int); // ff_isAsked_flag_set
 extern int quizQuestionAskedBitfield_get(u32); // ff_isAsked_flag_get
-
 extern void BKModel_getMeshCenter(BKModel *model, s32 mesh_id, s16 [3]); //! $a2 type unk
-
 extern void ability_setAllLearned(s32);  // set unlocked moves bitfield
 extern s32  ability_getAllLearned(void); // get unlocked moves bitfield
-
 extern s32  item_getCount(s32); // item count get
 extern void item_adjustByDiffWithoutHud(s32, s32); // item count set
-
 extern void func_8025A55C(s32, s32, s32);
-
 extern void func_80324CFC(f32, s16, s16);
-
 extern void  player_setTransformation(s32); // set transformation
-
-//typedef migrated to "inlcude/core2/code_C9E70.h"
 extern struct FF_StorageStruct *D_8037DCB8; 
 
 /* .data */
@@ -792,7 +784,7 @@ void func_8038D394(void)
     );
 }
 
-void func_8038D3F0(s32 a0, s32 a1)
+void func_8038D3F0(s32 a0, s8 a1)
 {
     if (a1 == -2)
     {
